requirements.txt
Django
djangorestframework
python-stdnum
Unidecode
openpyxl-dictreader
django-ninja
